os-observer/requirements.txt
pandas
python-dotenv
selenium
supabase